perf(windowing): perform histogram in worker
