+++ package.json
@@ -34,6 +34,7 @@
     "@velipso/polybool": "^2.0.11",
     "@vueuse/core": "^10.7.0",
     "@wdio/globals": "^9.7.1",
+    "comlink": "^4.4.1",
     "core-js": "3.22.5",
     "deep-equal": "^2.0.5",
     "dicomweb-client-typed": "^0.8.6",
