+++ src/utils/histogram.worker.ts
@@ -1,0 +1,8 @@
+import * as Comlink from 'comlink';
+import { histogram } from '@/src/utils/histogram';
+
+export interface HistogramWorker {
+  histogram: typeof histogram;
+}
+
+Comlink.expose({ histogram });
